updated npm axios dependency for security
