+++ Dahshop/ClientApp/vue-project/src/components/items/FileReader.vue
@@ -10,91 +10,23 @@
     />
     <ul v-if="imageList.length > 0">
       <li v-for="(image, index) in imageList" :key="index">
-        <img class="imagePreview" id="imgs" :src="image" />
+        <img class="imagePreview" id="imgs" :src="image" @load="revokeDataURL(index)" />
       </li>
     </ul>
-    <p v-else>List of images empty</p>
+    <p v-else>No images selected</p>
   </section>
 </template>
 
 <script>
-import { ref } from "vue";
+import useImagePreview from '../../hooks/imagePreview.js';
 
 export default {
   setup() {
     
-    const imageList = ref([]);
+    const { imageList, setDataURL, revokeDataURL } = useImagePreview();
 
-    function setDataURL(event) {
-
-      // set image list to an empty array just in case
-      imageList.value = [];
-
-      console.log(imageList.value.length);
-
-      // Reference to the DOM input elements
-      const files = event.target.files;
-
-      // Check file api support and if files is empty
-      if (files && window.File && window.FileList && window.FileReader) {
-        // Loop through the list of images uploaded
-        for (let i = 0; i < files.length; i++) {
-          // set file to files current index
-          const file = files[i];
-
-          console.log(file);
-
-          // set the current index in imagelist to an object url
-          imageList.value.push(URL.createObjectURL(file));
-
-          //   // Revoke the file to free memory
-          // URL.revokeObjectURL(this.imageList[i]);
-        }
-      }
-
-      console.log(imageList.value);
-      console.log(imageList.value.length);
-    }
-
-    return { imageList, setDataURL };
+    return { imageList, setDataURL, revokeDataURL };
   },
-
-  // data() {
-  //   return {
-  //     imageList: [],
-  //   };
-  // },
-  // // updated: {},
-  // methods: {
-  //   setDataURL(event) {
-  //     // set image list to an empty array just in case
-  //     this.imageList = [];
-  //     console.log(this.imageList.length);
-
-  //     // Reference to the DOM input elements
-  //     const files = event.target.files;
-
-  //     // Check file api support and if files is empty
-  //     if (files && window.File && window.FileList && window.FileReader) {
-  //       // Loop through the list of images uploaded
-  //       for (let i = 0; i < files.length; i++) {
-  //         // set file to files current index
-  //         const file = files[i];
-
-  //         console.log(file);
-
-  //         // set the current index in imagelist to an object url
-  //         this.imageList.push(URL.createObjectURL(file));
-
-  //         //   // Revoke the file to free memory
-  //         // URL.revokeObjectURL(this.imageList[i]);
-  //       }
-  //     }
-
-  //     console.log(this.imageList);
-  //     console.log(this.imageList.length);
-  //   },
-  // },
 };
 </script>
