+++ Dahshop/ClientApp/vue-project/src/components/items/ItemForm.vue
@@ -187,7 +187,7 @@ export default {
 
 <style scoped>
 .form-control {
-  margin: 0.5rem 0;
+  margin: 1rem 0;
 }
 
 label {
